This revision is the base from where icmake 7.10 will be created.
This revision will be tagged as 7.10-size_t-repaired



git-svn-id: https://icmake.svn.sourceforge.net/svnroot/icmake/icmake/trunk@207 dcc949cb-2a1c-0410-b67a-e7e5ca830264

+++ icmake/comp/bnot.c
@@ -12,7 +12,7 @@ ESTRUC_ *bnot (ESTRUC_ *e)                   /* expression so far */
         return (e);
     }
 
-    if ((e->type & ~ALLTYPES) == e_const)   /* immediate value */
+    if ((e->type & (size_t)~ALLTYPES) == e_const)   /* immediate value */
         e->evalue = ~e->evalue;
     else
     {
