+++ icmake/comp/pushdead.c
@@ -10,7 +10,7 @@ static size_t
 void push_dead()                            /* dead_sp: lastused, dead[0] = 0 */
 {
     if (++dead_sp >= size)                  /* too few elements ? then 5 more */
-        dead = xrealloc(dead, (size += 5) * sizeof(size_t));
+        dead = xrealloc(dead, (size += 5) * sizeof(unsigned));
 
     dead[dead_sp] = dead[dead_sp - 1];      /* copy former element */
 }
