+++ icmake/exec/stack/local.c
@@ -1,6 +1,6 @@
 #include "stack.ih"
 
-VAR_ *local(INT16 index)
+VAR_ *local(int index)
 {
-    return stack + (INT16)bp + (INT16)0xc000 - index;
+    return stack + (int)bp + (INT16)0xc000 - index;
 }
