+++ icmake/un/funpush1.c
@@ -2,5 +2,6 @@
 
 void fun_push_1_jmp_end ()
 {
-    printf ("        push int 1, jmp [%s]\n", hexstring (curoffs + 2, 4));
+    printf ("        push int 1, jmp [%s]\n", 
+                    hexstring((size_t)(curoffs + 2), 4));
 }
